+++ src/collective/nitf/tests/test_upgrades.py
@@ -84,7 +84,7 @@ class to2000TestCase(UpgradeTestCaseBase):
     def test_upgrade_to_2000_registrations(self):
         version = self.setup.getLastVersionForProfile(self.profile_id)[0]
         self.assertTrue(version >= self.to_version)
-        self.assertEqual(self.total_steps, 6)
+        self.assertEqual(self.total_steps, 8)
 
     @unittest.skipIf(IS_PLONE_5, 'Test not supported under Plone 5')
     def test_character_counter_css_resources(self):
@@ -240,28 +240,50 @@ class to2000TestCase(UpgradeTestCaseBase):
         new_permissions = ('collective.nitf: Setup',)
         self.assertEqual(configlet.getPermissions(), new_permissions)
 
-    def test_update_behavior(self):
+    def test_update_behaviors(self):
         # check if the upgrade step is registered
-        title = u'Update behavior'
+        title = u'Update behaviors'
         step = self.get_upgrade_step(title)
         self.assertIsNotNone(step)
 
         # simulate state on previous version
+        from collective.nitf.upgrades.v2000 import BEHAVIORS_TO_ADD
+        from collective.nitf.upgrades.v2000 import BEHAVIORS_TO_REMOVE
         fti = getUtility(IDexterityFTI, name='collective.nitf.content')
-        behaviors = list(fti.behaviors)
-        behaviors.remove('plone.app.relationfield.behavior.IRelatedItems')
-        behaviors.remove('collective.nitf.behaviors.interfaces.ISection')
-        behaviors.append('plone.app.referenceablebehavior.referenceable.IReferenceable')
-        fti.behaviors = tuple(behaviors)
-        self.assertIn('plone.app.referenceablebehavior.referenceable.IReferenceable', fti.behaviors)
-        self.assertNotIn('plone.app.relationfield.behavior.IRelatedItems', fti.behaviors)
-        self.assertNotIn('collective.nitf.behaviors.interfaces.ISection', fti.behaviors)
+        fti.behaviors = tuple(
+            set(fti.behaviors) - BEHAVIORS_TO_ADD | BEHAVIORS_TO_REMOVE)
+        for b in BEHAVIORS_TO_ADD:
+            self.assertNotIn(b, fti.behaviors)
+        for b in BEHAVIORS_TO_REMOVE:
+            self.assertIn(b, fti.behaviors)
 
         # run the upgrade step to validate the update
         self.execute_upgrade_step(step)
-        self.assertNotIn('plone.app.referenceablebehavior.referenceable.IReferenceable', fti.behaviors)
+        self.assertNotIn(
+            'plone.app.referenceablebehavior.referenceable.IReferenceable', fti.behaviors)
         self.assertIn('plone.app.relationfield.behavior.IRelatedItems', fti.behaviors)
         self.assertIn('collective.nitf.behaviors.interfaces.ISection', fti.behaviors)
+
+    def test_reindex_news_articles(self):
+        # check if the upgrade step is registered
+        title = u'Reindex news articles'
+        step = self.get_upgrade_step(title)
+        self.assertIsNotNone(step)
+
+        with api.env.adopt_roles(['Manager']):
+            for i in xrange(0, 10):
+                api.content.create(self.portal, 'collective.nitf.content', str(i))
+
+        # break the catalog by deleting an object without notifying
+        self.portal._delObject('0', suppress_events=True)
+        self.assertNotIn('0', self.portal)
+        results = api.content.find(portal_type='collective.nitf.content')
+        self.assertEqual(len(results), 10)  # catalog unaffected
+
+        # run the upgrade step to validate the update
+        self.execute_upgrade_step(step)
+        results = api.content.find(portal_type='collective.nitf.content')
+        self.assertEqual(len(results), 10)  # no failure and catalog unaffected
 
 
 class to2001TestCase(UpgradeTestCaseBase):