+++ src/collective/nitf/tests/test_upgrades.py
@@ -2,6 +2,7 @@
 from collective.nitf.testing import INTEGRATION_TESTING
 from collective.nitf.testing import IS_PLONE_5
 from plone import api
+from plone.dexterity.interfaces import IDexterityFTI
 from plone.registry import field
 from plone.registry.interfaces import IRegistry
 from plone.registry.record import Record
@@ -16,6 +17,9 @@ class UpgradeTestCaseBase(unittest.TestCase):
 
     def setUp(self, from_version, to_version):
         self.portal = self.layer['portal']
+        self.request = self.layer['request']
+        self.request.set('test', True)  # avoid transaction commits on tests
+
         self.setup = self.portal['portal_setup']
         self.profile_id = u'collective.nitf:default'
         self.from_version = from_version
@@ -30,10 +34,9 @@ class UpgradeTestCaseBase(unittest.TestCase):
 
     def execute_upgrade_step(self, step):
         """Execute an upgrade step."""
-        request = self.layer['request']
-        request.form['profile_id'] = self.profile_id
-        request.form['upgrades'] = [step['id']]
-        self.setup.manage_doUpgrades(request=request)
+        self.request.form['profile_id'] = self.profile_id
+        self.request.form['upgrades'] = [step['id']]
+        self.setup.manage_doUpgrades(request=self.request)
 
     @property
     def total_steps(self):
@@ -83,7 +86,7 @@ class to2000TestCase(UpgradeTestCaseBase):
     def test_upgrade_to_2000_registrations(self):
         version = self.setup.getLastVersionForProfile(self.profile_id)[0]
         self.assertTrue(version >= self.to_version)
-        self.assertEqual(self.total_steps, 4)
+        self.assertEqual(self.total_steps, 8)
 
     @unittest.skipIf(IS_PLONE_5, 'Test not supported under Plone 5')
     def test_character_counter_css_resources(self):
@@ -238,6 +241,49 @@ class to2000TestCase(UpgradeTestCaseBase):
         configlet = cptool.getActionObject('Products/nitf')
         new_permissions = ('collective.nitf: Setup',)
         self.assertEqual(configlet.getPermissions(), new_permissions)
+
+    def test_update_behaviors(self):
+        # check if the upgrade step is registered
+        title = u'Update behaviors'
+        step = self.get_upgrade_step(title)
+        self.assertIsNotNone(step)
+
+        # simulate state on previous version
+        from collective.nitf.upgrades.v2000 import BEHAVIORS_TO_ADD
+        REFERENCEABLE = 'plone.app.referenceablebehavior.referenceable.IReferenceable'
+        fti = getUtility(IDexterityFTI, name='collective.nitf.content')
+        fti.behaviors = tuple(
+            set(fti.behaviors) - BEHAVIORS_TO_ADD | set([REFERENCEABLE]))
+        for b in BEHAVIORS_TO_ADD:
+            self.assertNotIn(b, fti.behaviors)
+        self.assertIn(REFERENCEABLE, fti.behaviors)
+
+        # run the upgrade step to validate the update
+        self.execute_upgrade_step(step)
+        self.assertIn('plone.app.relationfield.behavior.IRelatedItems', fti.behaviors)
+        self.assertIn('collective.nitf.behaviors.interfaces.ISection', fti.behaviors)
+        self.assertIn(REFERENCEABLE, fti.behaviors)  # should not be removed
+
+    def test_reindex_news_articles(self):
+        # check if the upgrade step is registered
+        title = u'Reindex news articles'
+        step = self.get_upgrade_step(title)
+        self.assertIsNotNone(step)
+
+        with api.env.adopt_roles(['Manager']):
+            for i in xrange(0, 10):
+                api.content.create(self.portal, 'collective.nitf.content', str(i))
+
+        # break the catalog by deleting an object without notifying
+        self.portal._delObject('0', suppress_events=True)
+        self.assertNotIn('0', self.portal)
+        results = api.content.find(portal_type='collective.nitf.content')
+        self.assertEqual(len(results), 10)  # catalog unaffected
+
+        # run the upgrade step to validate the update
+        self.execute_upgrade_step(step)
+        results = api.content.find(portal_type='collective.nitf.content')
+        self.assertEqual(len(results), 10)  # no failure and catalog unaffected
 
 
 class to2001TestCase(UpgradeTestCaseBase):