+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 2.1b2 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
-- Fix upgrade process between versions 1.0 and 2.0 (fixes `#198`_).
+- Fix upgrade process between versions 1.0 and 2.0;
+  check documentation on migration from 1.x to 2.x (closes `#198`_).
   [rodfersou, hvelarde]