+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 2.1b2 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Do not create a link to ``None`` on the tile.
+  [hvelarde]
+
 - Fix upgrade process between versions 1.0 and 2.0;
   check documentation on migration from 1.x to 2.x (closes `#198`_).
   [rodfersou, hvelarde]